+++ MB.Presentation/Areas/Administrator/Pages/CommentManagement/List.cshtml.cs
@@ -24,13 +24,13 @@ namespace MB.Presentation.Areas.Administrator.Pages.CommentManagement
             Comments = _commentApplication.GetList();
         }
 
-        public RedirectToPageResult OnPostConfirm(int id)
+        public RedirectToPageResult OnPostConfirm(long id)
         {
             _commentApplication.Confirm(id);
             return RedirectToPage("./List", new { Confirmed = "True" });
         }
 
-        public RedirectToPageResult OnPostCancel(int id)
+        public RedirectToPageResult OnPostCancel(long id)
         {
             _commentApplication.Cancel(id);
             return RedirectToPage("./List", new { Canceled = "True" });
